[REL] account_multi_store to V11. [ADD] Separate Views and Py files by models. [FIX] Upgrade views. [FIX] Update README.
