+++ account_multi_store/README.rst
@@ -1,3 +1,15 @@
+.. |company| replace:: ADHOC SA
+
+.. |company_logo| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-logo.png
+   :alt: ADHOC SA
+   :target: https://www.adhoc.com.ar
+
+.. |icon| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-icon.png
+
+.. image:: https://img.shields.io/badge/license-AGPL--3-blue.png
+   :target: https://www.gnu.org/licenses/agpl
+   :alt: License: AGPL-3
+
 ==========================
 Multi Store for Accounting
 ==========================
@@ -16,6 +28,8 @@ Related Linked Models:
 Installation
 ============
 
+To install this module, you need to:
+
 This module should be auto-installed
 
 Configuration
@@ -26,17 +40,16 @@ To configure this module, you need to:
 #. Go to Configuration / Companies / Stores and create Store Journals
 #. or you can select an store for existing journals
 
+Usage
+=====
+
+To use this module, you need to:
+
+#. Just usage.
+
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
-   :target: https://runbot.adhoc.com.ar/
-
-.. repo_id is available in https://github.com/OCA/maintainer-tools/blob/master/tools/repos_with_ids.txt
-.. branch is "8.0" for example
-
-Known issues / Roadmap
-======================
-
-* ...
+   :target: http://runbot.adhoc.com.ar/
 
 Bug Tracker
 ===========
@@ -52,19 +65,16 @@ Credits
 Images
 ------
 
-* ADHOC SA: `Icon <http://fotos.subefotos.com/83fed853c1e15a8023b86b2b22d6145bo.png>`_.
+* |company| |icon|
 
 Contributors
 ------------
 
-
 Maintainer
 ----------
 
-.. image:: http://fotos.subefotos.com/83fed853c1e15a8023b86b2b22d6145bo.png
-   :alt: Odoo Community Association
-   :target: https://www.adhoc.com.ar
+|company_logo|
 
-This module is maintained by the ADHOC SA.
+This module is maintained by the |company|.
 
 To contribute to this module, please visit https://www.adhoc.com.ar.
