+++ account_multi_store/models/__init__.py
@@ -2,5 +2,9 @@
 # For copyright and license notices, see __manifest__.py file in module root
 # directory
 ##############################################################################
-from . import account
+from . import account_invoice
+from . import account_journal
+from . import account_move_line
+from . import account_move
+from . import account_payment
 from . import res_store
