+++ account_multi_store/models/account_move_line.py
@@ -2,4 +2,14 @@
 # For copyright and license notices, see __manifest__.py file in module root
 # directory
 ##############################################################################
-from . import models
+from odoo import models, fields
+
+
+class AccountMoveLine(models.Model):
+    _inherit = 'account.move.line'
+
+    store_id = fields.Many2one(
+        related='move_id.store_id',
+        store=True,
+        readonly=True,
+    )
