+++ account_multi_store/security/multi_store_security.xml
@@ -1,14 +1,12 @@
 <?xml version="1.0"?>
-<odoo>
-    <data noupdate="1">
-
+<odoo noupdate="1">
     <record id="account_journal_store_rule" model="ir.rule">
         <field name="name">Journal multi-store</field>
         <field name="model_id" ref="account.model_account_journal"/>
         <field name="global" eval="True"/>
         <field name="domain_force">['|',('store_id','=',False),('store_id','child_of',[user.store_id.id])]</field>
     </record>
-    
+
     <record id="account_move_store_rule" model="ir.rule">
         <field name="name">Account Entry</field>
         <field name="model_id" ref="account.model_account_move"/>
@@ -21,7 +19,7 @@
         <field name="name">Invoice multi-store</field>
         <field name="model_id" ref="account.model_account_invoice"/>
         <field name="global" eval="True"/>
-        <field name="perm_read" eval="False"/>        
+        <field name="perm_read" eval="False"/>
         <field name="domain_force">['|',('store_id','=',False),('store_id','child_of',[user.store_id.id])]</field>
     </record>
 
@@ -29,10 +27,8 @@
         <field name="name">Pyment multi-store</field>
         <field name="model_id" ref="account.model_account_payment"/>
         <field name="global" eval="True"/>
-        <field name="perm_read" eval="False"/>        
+        <field name="perm_read" eval="False"/>
         <field name="domain_force">['|',('store_id','=',False),('store_id','child_of',[user.store_id.id])]</field>
     </record>
-
-    </data>
 </odoo>
