+++ account_multi_store/views/account_invoice_views.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0"?>
+<odoo>
+    <!-- Invoice Search -->
+    <record id="view_account_invoice_filter" model="ir.ui.view">
+        <field name="name">account.invoice.select</field>
+        <field name="model">account.invoice</field>
+        <field name="inherit_id" ref="account.view_account_invoice_filter"/>
+        <field name="arch" type="xml">
+            <field name="journal_id" position="after">
+                <field name="store_id" groups="base_multi_store.group_multi_store"/>
+            </field>
+            <group position="inside">
+                <filter string="Store" context="{'group_by':'store_id'}" groups="base_multi_store.group_multi_store"/>
+            </group>
+        </field>
+    </record>
+</odoo>
