+++ account_multi_store/views/account_move_views.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0"?>
+<odoo>
+    <!-- Move Search -->
+    <record id="view_account_move_filter" model="ir.ui.view">
+        <field name="name">account.move.select</field>
+        <field name="model">account.move</field>
+        <field name="inherit_id" ref="account.view_account_move_filter"/>
+        <field name="arch" type="xml">
+            <field name="journal_id" position="after">
+                <field name="store_id" groups="base_multi_store.group_multi_store"/>
+            </field>
+            <group>
+                <filter string="Store" context="{'group_by':'store_id'}" groups="base_multi_store.group_multi_store"/>
+            </group>
+        </field>
+    </record>
+</odoo>
