+++ account_multi_store/views/account_payment_views.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0"?>
+<odoo>
+    <!-- Customer Voucher Search -->
+    <record id="view_account_payment_search" model="ir.ui.view">
+        <field name="name">account.payment.select</field>
+        <field name="model">account.payment</field>
+        <field name="inherit_id" ref="account.view_account_payment_search"/>
+        <field name="arch" type="xml">
+            <field name="journal_id" position="after">
+                <field name="store_id" groups="base_multi_store.group_multi_store"/>
+            </field>
+            <search>
+                <filter string="Store" context="{'group_by':'store_id'}" groups="base_multi_store.group_multi_store"/>
+            </search>
+        </field>
+    </record>
+</odoo>
