--- account_multi_store/views/account_view.xml
@@ -1,90 +1,0 @@
-<?xml version="1.0"?>
-<odoo>
-    <data>
-        <!-- Journal Form -->
-        <record id="view_account_journal_form" model="ir.ui.view">
-            <field name="name">account.journal.form</field>
-            <field name="model">account.journal</field>
-            <field name="inherit_id" ref="account.view_account_journal_form"/>
-            <field name="arch" type="xml">
-                <field name="company_id" position="after">
-                    <field name="store_id" groups="base_multi_store.group_multi_store"/>
-                </field>
-            </field>
-        </record>
-
-        <!-- Journal Search -->
-        <record id="view_account_journal_search" model="ir.ui.view">
-            <field name="name">account.journal.search</field>
-            <field name="model">account.journal</field>
-            <field name="inherit_id" ref="account.view_account_journal_search"/>
-            <field name="arch" type="xml">
-                <search>
-                    <field name="store_id" groups="base_multi_store.group_multi_store"/>
-                    <filter string="Store" context="{'group_by':'store_id'}" groups="base_multi_store.group_multi_store"/>
-                </search>
-            </field>
-        </record>
-
-        <!-- Invoice Search -->
-        <record id="view_account_invoice_filter" model="ir.ui.view">
-            <field name="name">account.invoice.select</field>
-            <field name="model">account.invoice</field>
-            <field name="inherit_id" ref="account.view_account_invoice_filter"/>
-            <field name="arch" type="xml">
-                <field name="journal_id" position="after">
-                    <field name="store_id" groups="base_multi_store.group_multi_store"/>
-                </field>
-                <group position="inside">
-                    <filter string="Store" context="{'group_by':'store_id'}" groups="base_multi_store.group_multi_store"/>
-                </group>
-            </field>
-        </record>
-
-        <!-- Customer Voucher Search -->
-        <record id="view_account_payment_search" model="ir.ui.view">
-            <field name="name">account.payment.select</field>
-            <field name="model">account.payment</field>
-            <field name="inherit_id" ref="account.view_account_payment_search"/>
-            <field name="arch" type="xml">
-                <field name="journal_id" position="after">
-                    <field name="store_id" groups="base_multi_store.group_multi_store"/>
-                </field>
-                <search>
-                    <filter string="Store" context="{'group_by':'store_id'}" groups="base_multi_store.group_multi_store"/>
-                </search>
-            </field>
-        </record>
-        
-        <!-- Move Search -->
-        <record id="view_account_move_filter" model="ir.ui.view">
-            <field name="name">account.move.select</field>
-            <field name="model">account.move</field>
-            <field name="inherit_id" ref="account.view_account_move_filter"/>
-            <field name="arch" type="xml">
-                <field name="journal_id" position="after">
-                    <field name="store_id" groups="base_multi_store.group_multi_store"/>
-                </field>
-                <group>
-                    <filter string="Store" context="{'group_by':'store_id'}" groups="base_multi_store.group_multi_store"/>
-                </group>
-            </field>
-        </record>
-        
-        <!-- Move Line Search -->
-        <record id="view_account_move_line_filter" model="ir.ui.view">
-            <field name="name">account.move.line.select</field>
-            <field name="model">account.move.line</field>
-            <field name="inherit_id" ref="account.view_account_move_line_filter"/>
-            <field name="arch" type="xml">
-                <field name="journal_id" position="after">
-                    <field name="store_id" groups="base_multi_store.group_multi_store"/>
-                </field>
-                <group>
-                    <filter string="Store" context="{'group_by':'store_id'}" groups="base_multi_store.group_multi_store"/>
-                </group>
-            </field>
-        </record>
-        
-    </data>
-</odoo>
