+++ account_multi_store/views/res_store_views.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0"?>
+<odoo>
+    <!-- Form -->
+    <record id="view_res_store_form" model="ir.ui.view">
+        <field name="name">res.store.form</field>
+        <field name="model">res.store</field>
+        <field name="inherit_id" ref="base_multi_store.view_res_store_form"/>
+        <field name="arch" type="xml">
+            <div name="buttons">
+                <button
+                    class="oe_inline oe_stat_button"
+                    name="%(account.action_account_journal_form)d"
+                    type="action"
+                    context="{'search_default_store_id': active_id}"
+                    icon="fa-money">
+                    <field name="journals_count" widget="statinfo"/>
+                </button>
+            </div>
+            <notebook>
+                <page string="Journals" name="journals">
+                    <field name="journal_ids"/>
+                </page>
+            </notebook>
+        </field>
+    </record>
+</odoo>
